Add readmodel db configuration types to common package
