+++ packages/commons/src/config/kafkaConfig.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+
+export const KafkaConfig = z
+  .object({
+    KAFKA_BROKERS: z.string(),
+    KAFKA_CLIENT_ID: z.string(),
+    KAFKA_GROUP_ID: z.string(),
+    KAFKA_DISABLE_AWS_IAM_AUTH: z.literal("true").optional(),
+  })
+  .transform((c) => ({
+    kafkaBrokers: c.KAFKA_BROKERS,
+    kafkaClientId: c.KAFKA_CLIENT_ID,
+    kafkaGroupId: c.KAFKA_GROUP_ID,
+    kafkaDisableAwsIamAuth: c.KAFKA_DISABLE_AWS_IAM_AUTH === "true",
+  }));
+export type KafkaConfig = z.infer<typeof KafkaConfig>;
